pyproject.toml
[project]
name = "simple-telegram-mcp"
version = "0.1.0"
description = "The simplest MCP server for interacting with Telegram"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name="Yaser Martinez Palenzuela", email="yaser.martinez@gmail.com" },
]
maintainers = [{ name = "Yaser Martinez Palenzuela", email = "yaser.martinez@gmail.com" }]
keywords = ["http", "mcp", "llm", "automation", "telegram"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "mcp>=1.1.3",
    "pydantic>=2.0.0",
    "python-dotenv",
    "telethon",
]

[project.scripts]
simple-telegram-mcp-server = "simple_telegram_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["pyright>=1.1.389", "ruff>=0.7.3"]
